- Hopefully this is the last restructuring of our source tree.
- include/ now matches src/ except that it may have a few extra
files. For example, we may create a engine/core.h which includes every
file in include/engine/core/ that we want to export. Now other modules
can include core.h without knowing what header files from
include/engine/core/ are being included under the hood.

- When doing a #include in a .h file, we should always uses the full
path from include/. For example, engine/core/tuple.h needs to specify
"engine/core/page.h" to include the page.h file. Otherwise, we are
forced to add an appropriate -Iengine/core flag to the compilation
path of any program that includes tuple.h. This can lead to conflicts
and collisions.

- When doing a #include in a .cpp file, feel free to avoid specifying
the full path of the .h and add another -I flag to the .cpp compile
options. This -I will be localized to a single Makefile, as opposed to
the -I requirement above which is forced on every Makefile.