** Integrating Mengzhi Wang's Shore TPC-C kit (fisrt commit) **
- The tpcc-kit has two sets of classes. The first one defines basic
metadata classes for accessing Shore files (like tables and indexes).
The second, defines the needed classes for running the tpcc benchmark
(like the tpcc-schema and tpcc-transactions).

- field, file
- table, table-scanner
- index, index-scanner
- tpcc-constants
- tpcc-tables
- tpcc-xct

Important notice:

This kit was written having a single threaded version of Shore in mind.
Thus, it is NOT thread-safe! We have to ensure that it runs correctly in
a multi-threaded environment. Some assertions have been inserted to places
where it was almost sure that it was not thread safe.