dora/range_action - Changing the implementation of range_action - Correcting the case where multiple keys need to be acquired on behalf of one action - Until now only the single record probes were correct
- The _up key value was not used
- Adding a vector of keys to be locked

DIRTY HACK WARNING:
- In the case of multiple keys, after the vector of keys is populated this list should *NOT
be modified. We are using the address of the elements in the vector. This normally
has undefined behavior if the vector is modified.

!!! BUG BUG BUG !!!
- Still we are not handling correctly the case when an action has to lock multiple
keys, and one of the keys it wants to lock it cannot be locked.
!!! BUG BUG BUG !!!


dora/base_action
- Adding a flag to indicate if this action is secondary or not.

dora/rvp
- Adding a function resize, which updates the counter of the actions to report.

dora/tm1
- Adding the implementation of the GET_SUB_NBR transaction, which uses a normal
(in SH-ared mode) secondary index probe and the secondary index has also the SID in
each entry. Then, the corresponding actions are populated and send to their executors.
The executor does the direct record access through the RID.