hacks to control vtune within shore-mt
I integrated the trick Ippokratis sent to pause and resume Vtune within Shore-MT. To enable it you have to configure shore-kits with --enable-vtune option.

Right now it pauses vtune at the beginning of shore-kits main function. Whenever 'measure' command is run it resumes VTune at the beginning of the measurement and later pauses it again at the end of that measurement. I assumed this scenario would be the most common profiling case for us so I only instrumented the measure command but related code changes are in shore_kits.cpp, so if you want to change pause/resume locations (search for CFG_VTUNE).

I also added a slighly modified version of the powerrun script (powerwrapper_vtune.sh) so that we can automate vtune profiling.

If you are going to use it for the first time for one of your papers please acknowledge Rene Mueller and Ippokratis (unless he is already a paper author :P)

cheers!